+++ bot/src/main/java/edu/java/bot/ListCommand.java
@@ -1,4 +1,4 @@
-package edu.java.bot.command;
+package edu.java.bot;
 
 import com.pengrad.telegrambot.model.Update;
 import com.pengrad.telegrambot.model.request.ParseMode;
